fix: rename debug ui component and optimize character list rendering
