+++ apps/client/src/game/CharacterActor.tsx
@@ -3,17 +3,22 @@ import type { Vector } from "@mp/math";
 import { EngineContext, Pixi } from "@mp/pixi/solid";
 import type { Character } from "@mp/server";
 import type { AreaResource } from "@mp/data";
+import type { Accessor } from "solid-js";
 import { createEffect, useContext } from "solid-js";
 import { Interpolator } from "./Interpolator";
 
-export function CharacterActor(props: { char: Character; area: AreaResource }) {
+export function CharacterActor(props: {
+  char: Accessor<Character>;
+  area: AreaResource;
+}) {
   const engine = useContext(EngineContext);
   const gfx = new CharacterGraphics();
   const lerp = new Interpolator(gfx);
 
   createEffect(() => {
     const { tiled } = props.area;
-    const { path, coords, speed } = props.char;
+    const { path, coords, speed } = props.char();
+
     gfx.update(tiled.tileSize);
     lerp.configure(tiled.tileCoordToWorld(coords), {
       path: path?.map(tiled.tileCoordToWorld) ?? [],
